feat(programmatic): update loading programmatic
