--- packages/oruga/src/components/loading/types.ts
@@ -1,5 +1,0 @@
-import type { ComponentProps } from "vue-component-type-helpers";
-
-import Loading from "./Loading.vue";
-
-export type LoadingProps = ComponentProps<typeof Loading>;
