+++ packages/oruga/src/composables/index.ts
@@ -5,7 +5,6 @@ export * from "./useMatchMedia";
 export * from "./useInputHandler";
 export * from "./useDebounce";
 export * from "./useParentProvider";
-export * from "./useProgrammatic";
 export * from "./useClickOutside";
 export * from "./useScrollingParent";
 export * from "./useObjectMap";
